Moved ternaries into data-structures package
