+++ src/app/pages/malbolge/models/data-structures/ternary.model.spec.ts
