+++ src/app/pages/malbolge/models/data-structures/ternary.model.ts
